add sqLite database with user authentication system
made a server dir with express for the backend added sqlite db for users with auth using JWT tokens and bycript for password hashing with 10 salte rounds for hasing. added a defult admin when the db is made for the first time. and added API for CURD for users and protected API routs